+++ packages/core/src/components/nav-menu/index.ts
@@ -1,1 +1,2 @@
 export { default as VfNavMenu } from './VfNavMenu.vue';
+export type { VfNavMenuProps } from '@/types/components';