+++ packages/core/src/entries/nav-menu.ts
@@ -1,3 +1,4 @@
 import '../styles/entries/nav-menu.css';
 
 export { VfNavMenu } from '../components/nav-menu';
+export type { VfNavMenuProps } from '../components/nav-menu';